requirements.txt
PyYAML==3.13
botocore==1.12.91
boto3==1.9.91
s3transfer==0.2.0
Flask==1.0.2
